ETL_OpenAI/requirements.txt
pandas
mysql-connector-python
openai==0.28
python-dotenv